requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
sqlalchemy
psycopg[binary]
psycopg2-binary
paddlepaddle==3.2.2
paddleocr
opencv-python
numpy
rapidfuzz